avs2: CB2 fix mosaic problems after seek operation. [1/1]
PD#SWPL-169420

Problem:
Because the PrevPicDistanceLsb variable is not updated in time when
decoding the error frame, the reference frame information is updated
incorrectly, which will cause abnormal decoding of subsequent decoded
frame reference messages.

Solution:
Update the PrevPicDistanceLsb variable in time when decoding the header
of each frame to ensure that the information in the reference
relationship is normal.

Verify:
S7D

Change-Id: Ifd18223195dad10508f38e871f92a6a40d97d8b1
Signed-off-by: gan.zhang <gan.zhang@amlogic.com>